Update godwoken scripts

+++ c/generator/secp256k1_helper.h
@@ -36,7 +36,7 @@ void secp256k1_default_error_callback_fn(const char* str, void* data) {
 int ckb_secp256k1_custom_verify_only_initialize(gw_context_t* ctx,
                                                 secp256k1_context* context,
                                                 void* data) {
-  uint32_t len = CKB_SECP256K1_DATA_SIZE;
+  uint64_t len = CKB_SECP256K1_DATA_SIZE;
   int ret =
       ctx->sys_load_data(ctx, ckb_secp256k1_data_hash, &len, 0, (uint8_t*)data);
   if (ret != 0 || len != CKB_SECP256K1_DATA_SIZE) {
